api/requirements.txt
fastapi
uvicorn
tensorflow-cpu
pandas
numpy
scikit-learn
h5py
pydantic
huggingface-hub